Gather the otp reminder to one file
