+++ src/constants/index.ts
@@ -7,9 +7,3 @@ export const FONTS = {
 export const GRAY_4 = '#F4F5F6'
 /** @deprecated */
 export const WHITE = '#FFFFFF'
-
-export const OTP_REMINDER_MILLISECONDS = 7 * 24 * 60 * 60 * 1000
-export const OTP_REMINDER_STORE_NAME = 'app.edge.login'
-export const OTP_REMINDER_KEY_NAME_CREATED_AT = 'createdAt'
-export const OTP_REMINDER_KEY_NAME_LAST_OTP_CHECKED = 'lastOtpCheck'
-export const OTP_REMINDER_KEY_NAME_DONT_ASK = 'OtpDontAsk'
